Extract thread control logic into a new ThreadedPeriodicWorker class
Instead of letting BackpressureMonitor, SessionFlusher, and Metrics::Aggregator
handle their own threading logic, we can extract that into a new class
that can be reused by all of them.
