+++ sentry-ruby/lib/sentry/session_flusher.rb
@@ -1,19 +1,15 @@
 # frozen_string_literal: true
 
 module Sentry
-  class SessionFlusher
-    include LoggingHelper
-
+  class SessionFlusher < ThreadedPeriodicWorker
     FLUSH_INTERVAL = 60
 
     def initialize(configuration, client)
-      @thread = nil
-      @exited = false
+      super(configuration.logger, FLUSH_INTERVAL)
       @client = client
       @pending_aggregates = {}
       @release = configuration.release
       @environment = configuration.environment
-      @logger = configuration.logger
 
       log_debug("[Sessions] Sessions won't be captured without a valid release") unless @release
     end
@@ -25,28 +21,16 @@ module Sentry
       @pending_aggregates = {}
     end
 
+    alias_method :run, :flush
+
     def add_session(session)
-      return if @exited
       return unless @release
 
-      begin
-        ensure_thread
-      rescue ThreadError
-        log_debug("Session flusher thread creation failed")
-        @exited = true
-        return
-      end
+      return unless ensure_thread
 
       return unless Session::AGGREGATE_STATUSES.include?(session.status)
       @pending_aggregates[session.aggregation_key] ||= init_aggregates(session.aggregation_key)
       @pending_aggregates[session.aggregation_key][session.status] += 1
-    end
-
-    def kill
-      log_debug("Killing session flusher")
-
-      @exited = true
-      @thread&.kill
     end
 
     private
@@ -69,17 +53,6 @@ module Sentry
 
     def attrs
       { release: @release, environment: @environment }
-    end
-
-    def ensure_thread
-      return if @thread&.alive?
-
-      @thread = Thread.new do
-        loop do
-          sleep(FLUSH_INTERVAL)
-          flush
-        end
-      end
     end
   end
 end
